LOVE HURTS!!!!!!!!!!!!!!!
I remember when I worked on the calendar implementation for _weeks_. And keep in mind, this is when I was still in school, so that means that it was late nights and early mornings for the whole thing to get implemented. Ah, who am I kidding, just late nights, there were _hardly_ mornings, and **never** early.

I slaved over the Ruby that I didn't have a damn clue about, and finally got a working implementation that incorporated all of the features that I wanted it to have. Now, it's time to switch to a more efficient alternative.

Incorporating this with our nextcloud instance is _certainly_ going to be more organized, and a better overall way of dealing with scheduling, but I'll be damned if I'm not a little miffed that my custom dynamically-formatted markdown-converted-to-html table is gone as a result. It's almost as if it didn't matter.

And yet, it's lead to bigger and better things...

And also smaller things... The Rakefile is 50% smaller after taking out the Calendar code. Go figure.